+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Add `ContainerCreateOptsBuilder::health_check_on_failure_action`
 * Add `ContainerCreateOptsBuilder::hostusers`
 * Add `ContainerCreateOptsBuilder::image_os`
+* Add `ContainerCreateOptsBuilder::image_variant`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync